Move fragmenter tests to their own module to make triaging panics easier.
The traige process filters out code in "_test.rs" files.
